requirements.txt
black==22.8.0
cffi==1.15.1
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pyheif @ git+https://github.com/david-poirier-csn/pyheif.git@721ef2c17347ccde7700d35e7d3b088d14fbae23
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.3.0
